Decouple web upload strategy
